Move initial recompilation to DrR loading time
This ensures that the splash screen does not disappear while the scripts are being
compiled, before the DrR frame shows up.

If the user later clicks on 'Recompile', after DrR has started, a frame% (not a dialog,
no need to be blocking) is shown to display a waiting message while the scripts are
being recompiled.