First attempt at hooks
A hook is similar to a script function, except that
i) there is no menu item associated with the script,
ii) the hook function is called automatically on certain events
such as file load, tab close, etc.
iii) the identifier of the hook function determines on what event
the hook function is called, e.g., `after-load-file` or `on-tab-close`.

The change is entirely backward compatible.

A user hook is defined similarly to a script function using the
`define-hook` form.
There is also no need to restart DrRacket for hooks to be activated
or deactived --- a simple click on `Scripts|Manage|Recompile scripts`
is enough.

There is no need for the user to use `define/augment`, by contrast to
writing a DrRacket plugin.

See the updated docs for more info.